# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weight-entries:
    get:
      operationId: listWeightEntries
      tags: [weight]
      summary: List weight entries ordered by date
      responses:
        "200":
          description: Weight entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightEntry"
    post:
      operationId: createWeightEntry
      tags: [weight]
      summary: Log a weight entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightEntry"
  /weight-entries/{id}:
    delete:
      operationId: deleteWeightEntry
      tags: [weight]
      summary: Delete a weight entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /measurements:
    get:
      operationId: listMeasurements
      tags: [weight]
      summary: List body area measurements
      responses:
        "200":
          description: Measurements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Measurement"
    post:
      operationId: createMeasurement
      tags: [weight]
      summary: Log a body area measurement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasurementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Measurement"
  /measurements/{id}:
    delete:
      operationId: deleteMeasurement
      tags: [weight]
      summary: Delete a measurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /goal:
    get:
      operationId: getGoal
      tags: [weight]
      summary: Get the active weight goal
      responses:
        "200":
          description: Goal (may be empty object defaults)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    put:
      operationId: setGoal
      tags: [weight]
      summary: Set or update the weight goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
components:
  schemas:
...
    WeightEntry:
      type: object
      required: [id, date, weightLbs]
      properties:
        id: { type: integer }
        date: { type: string }
        weightLbs: { type: number }
        note: { type: ["string", "null"] }
    WeightEntryInput:
      type: object
      required: [date, weightLbs]
      properties:
        date: { type: string }
        weightLbs: { type: number }
        note: { type: string }
    Measurement:
      type: object
      required: [id, date, area, valueInches]
      properties:
        id: { type: integer }
        date: { type: string }
        area: { type: string }
        valueInches: { type: number }
    MeasurementInput:
      type: object
      required: [date, area, valueInches]
      properties:
        date: { type: string }
        area: { type: string, minLength: 1 }
        valueInches: { type: number }
    Goal:
      type: object
      required: [id]
      properties:
        id: { type: integer }
        startWeightLbs: { type: ["number", "null"] }
        goalWeightLbs: { type: ["number", "null"] }
        targetDate: { type: ["string", "null"] }
        dailyCalorieTarget: { type: ["integer", "null"] }
    GoalInput:
      type: object
      properties:
        startWeightLbs: { type: number }
        goalWeightLbs: { type: number }
        targetDate: { type: string }
        dailyCalorieTarget: { type: integer }
...